requirements.txt
streamlit
tensorflow-cpu==2.8.0
tensorflow-hub==0.12.0
protobuf==3.20.0
pillow
numpy